+++ flashinfer/grouped_mm/__init__.py
@@ -1,10 +1,12 @@
-"""cuDNN-backed Grouped Matrix Multiplication (MoE Grouped GEMM).
+"""Grouped Matrix Multiplication (MoE Grouped GEMM).
 
 Provides grouped variants of the dense ``mm_*`` GEMM APIs, where each
 expert in a Mixture-of-Experts layer has its own weight matrix and tokens
 are routed to experts via ``m_indptr``.
 
-All implementations use cuDNN's ``moe_grouped_matmul`` as the backend.
+The public API is backend-agnostic and lives in :mod:`.core`.  Each
+backend has its own subpackage (e.g. :mod:`.cudnn`); future backends
+should follow the same convention.
 """
 
 from .core import grouped_mm_bf16 as grouped_mm_bf16
